# Fix degraded performance in filebuf